上传文件结束 404页面

+++ app/Http/Controllers/FilesController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\File;
+use \AetherUpload\RedisHandler as FileHsah;
+use Illuminate\Support\Facades\Storage;
 
 class FilesController extends Controller
 {
@@ -34,17 +36,27 @@ class FilesController extends Controller
     public function update(Request $request, File $file)
     {
         $file = $file->where('id',$request->id)->first();
-        if($request->logined == '1'){
+        if($request->logined == 'true'){
             $file->logined = true;
         }else{
             $file->logined = false;
         }
+        $file->save();
         return ['success' => true,'msg'=>'文件下载权限已更新。'];
     }
     public function destroy(Request $request, File $file)
     {
+        $res = ['success' => false,'msg'=>'文件不存在。'];
         $file = $file->where('id',$request->id)->first();
-        $file->delete();
-        return ['success' => true,'msg'=>'文件删除成功。'];
+        if(!$file){
+            return $res;
+        }
+        $result =  Storage::delete('aetherupload/'.$file->path);
+        if($result){
+            FileHsah::deleteOneHash($file->hash);
+            $file->delete();
+            $res = ['success' => true,'msg'=>'文件删除成功。'];
+        }
+        return $res;
     }
 }
